pull the app::parse::Bytes type out into its own module and rexport into app
