+++ ffi/dnp3-ffi/src/handler.rs
@@ -1,7 +1,6 @@
 use crate::ffi;
 use dnp3::app::header::{Iin1, Iin2, ResponseFunction, ResponseHeader};
 use dnp3::app::measurement::*;
-use dnp3::app::parse::bytes::Bytes;
 use dnp3::app::*;
 use dnp3::master::handle::{HeaderInfo, ReadHandler};
